Add `Gui::update`
